Changed aggregate initialization to be able to drop more aggregate expressions that can be derived from the start.

+++ solveridsmt/aggs/AggTypes.h
@@ -18,7 +18,7 @@ using namespace boost;
 
 #include "BigInteger.hh"
 #include "BigIntegerUtils.hh"
-typedef BigInteger Weight;
+typedef int Weight;
 
 namespace Aggrs{
 class WLV;
